+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - @aofl/cli - serve command
 - @aofl/templating-plugin - eslint-disable routes-config-loader
 - @aofl/webcomponent-css-loader - added cache option
+- @aofl/webcomponent-css-loader - added whitelist option
 - @aofl/store - auto generates setter mutations based on init() mutation
 - @aofl/dom-scope - added README.md
 - @aofl/dom-scope - added cache option